1.1.0-pre: Rearranged FW panel layout
This commit adds a new, more comfy feel to the firmware commands layout,
replacing all the ugly buttons with a nice list view and minimal input fields.

Some previously-committed changes to the UI were reverted on accident;
this will be investigated later.

Menus have been added to support some advanced USB & COM capabilities,
but the functionality has not been implemented yet.

Some of the baud-rate support has been tweaked, so it may not be able to run
at >9600 baud.  Some Bluetooth drivers have a problem closing and then
quickly re-opening a port again; this is currently being worked.  Right now,
the port is not being closed so it stays open in order to prevent problems
transmitting messages over Bluetooth.

Due to progress with hardware testing, the "Update the firmware on your
boards here" section should soon be activated!