Massive speed-up to serial code
The way in which messages are synthesized and written to the boards has changed drastically, and now the software is so efficient that the firmware had to be modified in order to keep up.  This also possibly includes some of the early code for TrueType font support and custom board layouts, which is still highly experimental at this time.  Mac OSX has been observed not to do some of the firmware operations correctly, so stay tuned for enhancements to that area in the next commit.