Improve "compressChipIds" function
The UI now allows the user to enter a "gap" so that boards aren't just reassigned sequential values.  That is, with a gap of 4, your boards will be reassigned as 0, 4, 8, 12, 16, ...  The logic in the compressChipIds function has been modified to support the UI changes, and now works a bit better than it did before as well.

There are also a couple other minor changes.