Version 1.2 for Windows and Linux-64
Renamed many things from LEDgoes to BriteBlox, and made other minor UI adjustments.

SerialWriteEx & SerialReadEx now return a boolean value to indicate success or failure.  This way, if the function fails, the calling function can quit running the loop or take other action so as not to waste a bunch of time.

Multiple rows are now allowed and are easy to configure in the PC Tools.  The _updateBoards() function will make the appropriate accommodations.  Also, to support this, the "Custom Layout..." button has now been fixed so that the Custom Layout Editor is now shown.  That feature, though, is still a big work in progress.

compressChipIDs now leverages from balanceChipIDs in order to help rearrange the chip & board IDs if desired by the user.  balanceChipIDs() evenly spaces the IDs out in case there is any "sag" due to non-ideal voltage drops, and compressChipIDs() sets balanced IDs to the requested starting value & offset per board, such as [0,1,2,3], [18,20,22], etc...

"About" menu item now actually does something. :-P

Added an experimental change to the functions that control baud rate.  Now when you change the baud rate in the dropdown, the panels instantly change their baud rate.  This no longer happens upon beginning to scroll a message.  If you have thoughts about this, please let us know.

Rearranged various statements in showFwHelp() to be in the same order all the time.