New features this weekend
Added a rough capability to read RSS feeds, particularly to reach out to NASDAQ to get stock quotes for publicly-traded companies in the US.

Improved the efficiency of the makeMessageString() function so that it no longer parses the XML tree twice for each message.

Added the concept of having updateMessages() wait for asynchronous tasks to finish before the message string is re-computed.  This will help update data sources involving pull-based notifications (like RSS feeds but unlike Twitter).