pushMsg will now accept a static message index number for an incoming message.  This means something like the stock ticker can always update message #1 in the queue.  (We don't want it showing old quotes, do we? ;)
Fixed a bug relating to editing selected messages when selecting them with the mouse vs. keyboard.

Added the "Replace" feature so we can finally replace messages in the queue.

Replaced the RSS panel's initialization code in the main file to call the "RSS" library's init_ui() function.  This sets the stage for the standard way to initialize UI panels for this application in the future, assuming I don't rewrite it in C++ in the meantime :-P.

Fixed some of the "hackiness" around the stock quote/RSS module.