Add files via upload
This adds in some very early experimental support for Windows fonts on BriteBlox displays.  The configuration I would normally run this in would be a 2-panel-tall matrix so as to display Arial at something like 14 or 16pt.  In such a double tall setup, this might work best in conjunction with the Panel Editor so you can specify which BriteBlox Displayboard addresses you have where.